% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_palindrome.R
\name{rm_palindrome}
\alias{rm_palindrome}
\title{Remove Palindromic SNPs}
\usage{
rm_palindrome(testData, t)
}
\arguments{
\item{testData}{data.frame or data.table containing "reference_allele" and "other_allele" columns, each column contains single character (either A, T, G, or C). Note that the function uses rs_number to remove the SNPs from the original data.frame, making "rs_number" a required column.}

\item{t}{Threshold value >0 & <=0.5}
}
\value{
data.table or data.frame with removed palindromic SNPs within 0.5 +/- threshold t (note that function must be called and assigned to a data.frame, see examples)
}
\description{
Detect SNPs palindromic with palindromic alleles and remove them if their allele frequency is within a specific threshold around the allele frequency of 0.5.
}
\examples{
result_df <- rm_palindrome(input_df,t);
result_df <- rm_palindrome(myData,0.2);
}
